refactor: access token에서 userId 뺴는 함수 추출
