+++ frontend/techpick/src/utils/index.ts
@@ -21,3 +21,4 @@ export { getItemFromLocalStorage } from './getItemFromLocalStorage';
 export { setItemToLocalStorage } from './setItemToLocalStorage';
 export { isMacOS } from './isMacOS';
 export { isLoginUser } from './isLoginUser';
+export { getUserIdForServer } from './getUserIdForServer';
